Start cli refactor
