+++ kanidm_tools/src/cli/account.rs
@@ -1,12 +1,12 @@
 
 #[derive(Debug, StructOpt)]
-struct AccountCommonOpt {
+pub struct AccountCommonOpt {
     #[structopt()]
     account_id: String,
 }
 
 #[derive(Debug, StructOpt)]
-struct AccountCredentialSet {
+pub struct AccountCredentialSet {
     #[structopt(flatten)]
     aopts: AccountCommonOpt,
     #[structopt()]
@@ -16,7 +16,7 @@ struct AccountCredentialSet {
 }
 
 #[derive(Debug, StructOpt)]
-struct AccountNamedOpt {
+pub struct AccountNamedOpt {
     #[structopt(flatten)]
     aopts: AccountCommonOpt,
     #[structopt(flatten)]
@@ -24,7 +24,7 @@ struct AccountNamedOpt {
 }
 
 #[derive(Debug, StructOpt)]
-struct AccountNamedTagOpt {
+pub struct AccountNamedTagOpt {
     #[structopt(flatten)]
     aopts: AccountCommonOpt,
     #[structopt(flatten)]
@@ -34,7 +34,7 @@ struct AccountNamedTagOpt {
 }
 
 #[derive(Debug, StructOpt)]
-struct AccountNamedTagPKOpt {
+pub struct AccountNamedTagPKOpt {
     #[structopt(flatten)]
     aopts: AccountCommonOpt,
     #[structopt(flatten)]
@@ -46,7 +46,7 @@ struct AccountNamedTagPKOpt {
 }
 
 #[derive(Debug, StructOpt)]
-struct AccountCreateOpt {
+pub struct AccountCreateOpt {
     #[structopt(flatten)]
     aopts: AccountCommonOpt,
     #[structopt(name = "display_name")]
@@ -56,7 +56,7 @@ struct AccountCreateOpt {
 }
 
 #[derive(Debug, StructOpt)]
-enum AccountCredential {
+pub enum AccountCredential {
     #[structopt(name = "set_password")]
     SetPassword(AccountCredentialSet),
     #[structopt(name = "generate_password")]
@@ -64,7 +64,7 @@ enum AccountCredential {
 }
 
 #[derive(Debug, StructOpt)]
-enum AccountRadius {
+pub enum AccountRadius {
     #[structopt(name = "show_secret")]
     Show(AccountNamedOpt),
     #[structopt(name = "generate_secret")]
@@ -74,7 +74,7 @@ enum AccountRadius {
 }
 
 #[derive(Debug, StructOpt)]
-struct AccountPosixOpt {
+pub struct AccountPosixOpt {
     #[structopt(flatten)]
     aopts: AccountCommonOpt,
     #[structopt(long = "gidnumber")]
@@ -86,7 +86,7 @@ struct AccountPosixOpt {
 }
 
 #[derive(Debug, StructOpt)]
-enum AccountPosix {
+pub enum AccountPosix {
     #[structopt(name = "show")]
     Show(AccountNamedOpt),
     #[structopt(name = "set")]
@@ -96,7 +96,7 @@ enum AccountPosix {
 }
 
 #[derive(Debug, StructOpt)]
-enum AccountSsh {
+pub enum AccountSsh {
     #[structopt(name = "list_publickeys")]
     List(AccountNamedOpt),
     #[structopt(name = "add_publickey")]
@@ -106,7 +106,7 @@ enum AccountSsh {
 }
 
 #[derive(Debug, StructOpt)]
-enum AccountOpt {
+pub enum AccountOpt {
     #[structopt(name = "credential")]
     Credential(AccountCredential),
     #[structopt(name = "radius")]
@@ -125,3 +125,157 @@ enum AccountOpt {
     Delete(AccountNamedOpt),
 }
 
+impl AccountOpt {
+    pub fn exec(&self) -> () {
+        match self {
+            // id/cred/primary/set
+            AccountOpt::Credential(acopt) => match acopt {
+                AccountCredential::SetPassword(acsopt) => {
+                    let client = acsopt.copt.to_client();
+                    let password = rpassword::prompt_password_stderr(
+                        format!("Enter new password for {}: ", acsopt.aopts.account_id).as_str(),
+                    )
+                    .unwrap();
+
+                    client
+                        .idm_account_primary_credential_set_password(
+                            acsopt.aopts.account_id.as_str(),
+                            password.as_str(),
+                        )
+                        .unwrap();
+                }
+                AccountCredential::GeneratePassword(acsopt) => {
+                    let client = acsopt.copt.to_client();
+
+                    let npw = client
+                        .idm_account_primary_credential_set_generated(
+                            acsopt.aopts.account_id.as_str(),
+                        )
+                        .unwrap();
+                    println!(
+                        "Generated password for {}: {}",
+                        acsopt.aopts.account_id, npw
+                    );
+                }
+            }, // end AccountOpt::Credential
+            AccountOpt::Radius(aropt) => match aropt {
+                AccountRadius::Show(aopt) => {
+                    let client = aopt.copt.to_client();
+
+                    let rcred = client
+                        .idm_account_radius_credential_get(aopt.aopts.account_id.as_str())
+                        .unwrap();
+
+                    match rcred {
+                        Some(s) => println!("Radius secret: {}", s),
+                        None => println!("NO Radius secret"),
+                    }
+                }
+                AccountRadius::Generate(aopt) => {
+                    let client = aopt.copt.to_client();
+                    client
+                        .idm_account_radius_credential_regenerate(aopt.aopts.account_id.as_str())
+                        .unwrap();
+                }
+                AccountRadius::Delete(aopt) => {
+                    let client = aopt.copt.to_client();
+                    client
+                        .idm_account_radius_credential_delete(aopt.aopts.account_id.as_str())
+                        .unwrap();
+                }
+            }, // end AccountOpt::Radius
+            AccountOpt::Posix(apopt) => match apopt {
+                AccountPosix::Show(aopt) => {
+                    let client = aopt.copt.to_client();
+                    let token = client
+                        .idm_account_unix_token_get(aopt.aopts.account_id.as_str())
+                        .unwrap();
+                    println!("{:?}", token);
+                }
+                AccountPosix::Set(aopt) => {
+                    let client = aopt.copt.to_client();
+                    client
+                        .idm_account_unix_extend(
+                            aopt.aopts.account_id.as_str(),
+                            aopt.gidnumber,
+                            aopt.shell.as_deref(),
+                        )
+                        .unwrap();
+                }
+                AccountPosix::SetPassword(aopt) => {
+                    let client = aopt.copt.to_client();
+                    let password =
+                        rpassword::prompt_password_stderr("Enter new unix (sudo) password: ")
+                            .unwrap();
+                    client
+                        .idm_account_unix_cred_put(
+                            aopt.aopts.account_id.as_str(),
+                            password.as_str(),
+                        )
+                        .unwrap();
+                }
+            }, // end AccountOpt::Posix
+            AccountOpt::Ssh(asopt) => match asopt {
+                AccountSsh::List(aopt) => {
+                    let client = aopt.copt.to_client();
+
+                    let pkeys = client
+                        .idm_account_get_ssh_pubkeys(aopt.aopts.account_id.as_str())
+                        .unwrap();
+
+                    for pkey in pkeys {
+                        println!("{}", pkey)
+                    }
+                }
+                AccountSsh::Add(aopt) => {
+                    let client = aopt.copt.to_client();
+                    client
+                        .idm_account_post_ssh_pubkey(
+                            aopt.aopts.account_id.as_str(),
+                            aopt.tag.as_str(),
+                            aopt.pubkey.as_str(),
+                        )
+                        .unwrap();
+                }
+                AccountSsh::Delete(aopt) => {
+                    let client = aopt.copt.to_client();
+                    client
+                        .idm_account_delete_ssh_pubkey(
+                            aopt.aopts.account_id.as_str(),
+                            aopt.tag.as_str(),
+                        )
+                        .unwrap();
+                }
+            }, // end AccountOpt::Ssh
+            AccountOpt::List(copt) => {
+                let client = copt.to_client();
+                let r = client.idm_account_list().unwrap();
+                for e in r {
+                    println!("{:?}", e);
+                }
+            }
+            AccountOpt::Get(aopt) => {
+                let client = aopt.copt.to_client();
+                let e = client
+                    .idm_account_get(aopt.aopts.account_id.as_str())
+                    .unwrap();
+                println!("{:?}", e);
+            }
+            AccountOpt::Delete(aopt) => {
+                let client = aopt.copt.to_client();
+                client
+                    .idm_account_delete(aopt.aopts.account_id.as_str())
+                    .unwrap();
+            }
+            AccountOpt::Create(acopt) => {
+                let client = acopt.copt.to_client();
+                client
+                    .idm_account_create(
+                        acopt.aopts.account_id.as_str(),
+                        acopt.display_name.as_str(),
+                    )
+                    .unwrap();
+            }
+        }
+    }
+}
